* dev:
  DOC: Update requirements
  STYLE: Interfaces to use flirt/regaladin for v2v reg + choose visualization viewer
  BUG: Create masks as np.uint8 in stack_mask_morphological_operations
  BUG: slice thickness option not passed for N4ITK script
  DOC: Update of target-stack-index help text
  ENH: Slice thicknesses of image stacks can be manually specified
  ENH: Allow dir_input_mc for evaluate_slice_residual_similarity
  DOC: Update README
  DOC: Update README

+++ niftymic/utilities/stack_mask_morphological_operations.py
@@ -108,7 +108,7 @@ class StackMaskMorphologicalOperations(object):
 
             shape = self._mask_sitk.GetSize()
             N_slices = shape[2]
-            nda_mask = np.zeros(shape[::-1], dtype=np.uint16)
+            nda_mask = np.zeros(shape[::-1], dtype=np.uint8)
 
             for i in range(0, N_slices):
                 slice_mask_sitk = self._mask_sitk[:,:,i:i+1]
